move message manager
